[510:sparkles:] Configurable Roles for RBAC Verification
Added the feature without much fuss, but found a ton of issues with the
existing unit tests for the config settings. Not entirely sure if the
current strategy for testing is very good. Created a follow-up ticket [529] to
revisit the test suite in the future.

Updated the docs as well to reflect the new exciting feature!