[538:sparkles:] Update Docs with Quickstart Guide
Quickstart Guide is 99% complete, but I ran into a chicken/egg issue with
referencing 'flask-ligand-example' project code in the docs. So a release
needs to be made before completion of docs.

I discovered while writing documentation that I was not utilizing 'create_app'
correctly for the 'flask-ligand-example' and it turns out I had a design flaw
in how I was registering Blueprints. Fixing the issue required a signature
change to 'create_app' that rippled to 'flask-ligand-example', thus causing
the circular reference issue.

Anywho, releasing a new version early!