[500:sparkles:] Add Auth0 Support
Updated settings configuration flow to allow for an empty string for
OIDC_REALM which allows configuring flask-ligand for use with Auth0!
Created documentation on how to setup Auth0 and configure flask-ligand
to get full RBAC support.

Discovered that the system for detecting missing required settings was
janky, so I refactored it to make more sense. Discovered a bunch of
minor bugs and documentation issues along the way that were fixed as
well.