[511:sparkles:] Fix OIDC IAM Provider Configuration
The OIDC_REALM setting is superfluous because the user should be providing a
valid URL to the OpenID Connect Discovery Mechanism which already incorporates
providing a realm for multi-tenant issuers like Keycloak. The OIDC_ISSUER_URL
should be changed to OIDC_DISCOVERY_URL for clarity's sake.

The docs should be updated providing basic information on how to determine the
OIDC_DISCOVERY_URL for the providers that are currently supported.