[543:robot:] Update Dependencies
The recent major version release of Flask-SQLAlchemy deprecated features
that flask-ligand depends on. The code was refactored to address the
upcoming breaking changes.

The circular dependency with the 'flask-ligand-example' library for building
docs became too cumbersome to work-around, so I gave up and just copied the
code directly into the quickstart guide.

Also, I bumped the version requirement for 'flask-smorest' since it was
recently updated.