[536:sparkles:] Generate OpenAPI Clients Offline
In order to support this feature well I created a new environment "cli"
which will allow the service to run without initializing Flask extensions.

I found a bunch of minor typos again that I fixed as part of this feature
because I'm lazy and don't want to create another commit.